# docs/daemon/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ADF Daemon API",
    "version": "0.1.0",
    "description": "Static OpenAPI contract for the local ADF daemon. The daemon is a headless ADF runtime for loading agents, starting and stopping them, sending chat turns, streaming events, inspecting runtime diagnostics, and resolving human-in-the-loop tasks. Every route except GET /health requires `Authorization: Bearer <token>` (the per-install token in <settings dir>/daemon-token, or ADF_DAEMON_TOKEN); requests with a disallowed Host (403 host_not_allowed), a foreign Origin or Sec-Fetch-Site cross-site/same-site (403 cross_origin) are refused, and no CORS headers are sent. Credential reads return metadata only."
  },
  "security": [{ "bearerAuth": [] }],
  "servers": [
    {
      "url": "http://127.0.0.1:7385",
      "description": "Default local daemon"
    }
  ],
...
  "paths": {
...
    "/templates": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "List agent templates new agents can be created from",
        "operationId": "listTemplates",
        "responses": {
          "200": {
            "description": "Templates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateList"
                }
              }
            }
          },
          "409": {
            "description": "Owner identity not ready; `identity` carries the status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      },
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Create a template: blank, or a duplicate of fromId (no identity or history carried)",
        "operationId": "createTemplate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "fromId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Owner identity not ready (`identity_not_ready`; `identity` carries the status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/templates/{id}": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "One template: summary, contents (config, seed files README.md / mind.md / soul.md, extra files) and whether it is the default",
        "operationId": "getTemplate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Template id (the file stem in the templates folder)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Owner identity not ready (`identity_not_ready`; `identity` carries the status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      },
      "patch": {
        "tags": [
          "Agents"
        ],
        "summary": "Rename a template and/or set its notes (description, warning; an empty string clears one). A rename moves the file: the response carries the new id",
        "operationId": "updateTemplate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Template id (the file stem in the templates folder)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "warning": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Owner identity not ready (`identity_not_ready`; `identity` carries the status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "Agents"
        ],
        "summary": "Delete a template: the file moves to <userData>/templates-trash (never a hard delete); the default falls back to standard",
        "operationId": "deleteTemplate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Template id (the file stem in the templates folder)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Owner identity not ready (`identity_not_ready`; `identity` carries the status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/templates/{id}/default": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Make this template the default for new agents",
        "operationId": "setDefaultTemplate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Template id (the file stem in the templates folder)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Owner identity not ready (`identity_not_ready`; `identity` carries the status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/templates/{id}/reset": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Reset a shipped template to the version the app ships (400 for a user template)",
        "operationId": "resetShippedTemplate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Template id (the file stem in the templates folder)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Owner identity not ready (`identity_not_ready`; `identity` carries the status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/templates/{id}/review": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "The agent review summary for a template (needsReview, reviewed, summary)",
        "operationId": "getTemplateReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Template id (the file stem in the templates folder)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Owner identity not ready (`identity_not_ready`; `identity` carries the status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/templates/{id}/review/accept": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Accept a template's review: claims the file with a fresh identity under this owner, auto-locks security fields, marks it reviewed. password for a password-protected file (400 password_required, 403 wrong_password)",
        "operationId": "acceptTemplateReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Template id (the file stem in the templates folder)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Owner identity not ready (`identity_not_ready`; `identity` carries the status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/templates/{id}/config": {
      "put": {
        "tags": [
          "Agents"
        ],
        "summary": "Replace a template's agent config (validated against the config schema)",
        "operationId": "setTemplateConfig",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Template id (the file stem in the templates folder)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "config"
                ],
                "properties": {
                  "config": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Owner identity not ready (`identity_not_ready`; `identity` carries the status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/templates/{id}/files": {
      "put": {
        "tags": [
          "Agents"
        ],
        "summary": "Write a template file: a seed file (README.md, mind.md, soul.md) or an extra file",
        "operationId": "setTemplateFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Template id (the file stem in the templates folder)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "path",
                  "content"
                ],
                "properties": {
                  "path": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Owner identity not ready (`identity_not_ready`; `identity` carries the status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "Agents"
        ],
        "summary": "Remove an extra file (seed files cannot be removed; clear their text instead)",
        "operationId": "removeTemplateFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Template id (the file stem in the templates folder)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Object"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Owner identity not ready (`identity_not_ready`; `identity` carries the status)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCreateErrorResponse"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-install token from <settings dir>/daemon-token (printed by `adf daemon token`), or ADF_DAEMON_TOKEN. Local adf clients send it automatically."
      }
    },
...
    "responses": {
...
      "Object": {
        "description": "JSON object",
        "content": {
          "application/json": {
            "schema": { "type": "object", "additionalProperties": true }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "Unavailable": {
        "description": "Subsystem not configured",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "Error": {
        "description": "Runtime error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
...
      "IdentityStatus": {
        "type": "object",
        "required": [
          "status",
          "ownerDid",
          "runtimeDid",
          "storage",
          "backupConfirmed",
          "passphraseRequired",
          "message"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "none",
              "locked",
              "restore-needed",
              "ready"
            ]
          },
          "ownerDid": {
            "type": [
              "string",
              "null"
            ]
          },
          "runtimeDid": {
            "type": [
              "string",
              "null"
            ],
            "description": "This daemon's own runtime DID (never Studio's)."
          },
          "storage": {
            "type": "string",
            "enum": [
              "keychain",
              "file"
            ]
          },
          "backupConfirmed": {
            "type": "boolean"
          },
          "passphraseRequired": {
            "type": "boolean",
            "description": "File storage not unlocked: create/restore/unlock take a passphrase."
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "TemplateList": {
        "type": "object",
        "required": [
          "templates",
          "defaultId"
        ],
        "properties": {
          "templates": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "name",
                "reviewed"
              ],
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "templateDescription": {
                  "type": "string"
                },
                "warning": {
                  "type": "string"
                },
                "shipped": {
                  "type": "string"
                },
                "reviewed": {
                  "type": "boolean"
                },
                "modelProvider": {
                  "type": "string"
                },
                "modelId": {
                  "type": "string"
                }
              },
              "additionalProperties": true
            }
          },
          "defaultId": {
            "type": "string"
          },
          "folder": {
            "type": "string"
          },
          "defaultDirectory": {
            "type": "string",
            "description": "Where POST /agents/create puts agents without `directory`."
          }
        }
      },
...
      "AgentCreateErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "identity": {
            "$ref": "#/components/schemas/IdentityStatus"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" }
        },
        "additionalProperties": true
      },
...
    }
  }
}